Refactoring of the Glutin port in preparation of the compositor refactoring.
